requirements.txt
pytesseract==0.3.10
Pillow==10.0.1
reportlab==4.0.4
opencv-python==4.8.1.78
numpy==1.24.3
streamlit==1.28.1
pdf2image==1.16.3
PyPDF2==3.0.1 